moved components to their own folder
