+++ app/portfolio/page.tsx
@@ -4,10 +4,10 @@ import React, { useState, useEffect, useMemo } from 'react'
 import { motion, AnimatePresence } from 'framer-motion'
 import { Button } from '@/components/ui/button'
 import { projects, type Project } from '@/lib/projects'
-import { CustomLightbox } from '@/components/custom-lightbox'
-import { ProjectCard } from '@/components/project-card'
-import { ViewMoreCard } from '@/components/view-more-card'
-import { PageHeroSection } from '@/components/page-hero-section'
+import { CustomLightbox } from '@/components/CustomLightbox'
+import { ProjectCard } from '@/components/ProjectCard'
+import { ViewMoreCard } from '@/components/ViewMoreCard'
+import { PageHeroSection } from '@/components/PageHeroSection'
 
 type CategoryType = 'all' | 'photography' | 'video-production' | 'weddings'
