+++ app/services/portrait/page.tsx
@@ -4,11 +4,11 @@ import { motion } from 'framer-motion'
 import Link from 'next/link'
 import { Button } from '@/components/ui/button'
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card'
-import { CategoryFilter } from '@/components/category-filter'
-import { PageHeroSection } from '@/components/page-hero-section'
+import { CategoryFilter } from '@/components/CategoryFilter'
+import { PageHeroSection } from '@/components/PageHeroSection'
 import { projects, type Project } from '@/lib/projects'
 import { useState, useMemo } from 'react'
-import { CustomLightbox } from '@/components/custom-lightbox'
+import { CustomLightbox } from '@/components/CustomLightbox'
 import { sectionLabels, portraitFilterCategories } from '@/data/constants'
 import { Heart, Users, Camera } from 'lucide-react'
