+++ app/services/wedding/page.tsx
@@ -4,10 +4,10 @@ import { motion } from 'framer-motion'
 import Link from 'next/link'
 import { Button } from '@/components/ui/button'
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card'
-import { CategoryFilter } from '@/components/category-filter'
-import { PageHeroSection } from '@/components/page-hero-section'
+import { CategoryFilter } from '@/components/CategoryFilter'
+import { PageHeroSection } from '@/components/PageHeroSection'
 import { useState, useMemo } from 'react'
-import { CustomLightbox } from '@/components/custom-lightbox'
+import { CustomLightbox } from '@/components/CustomLightbox'
 import { Heart, Star } from 'lucide-react'
 import { projects, type Project } from '@/lib/projects'
 import { weddingFilterCategories } from '@/data/constants'
