+++ components/CustomLightbox/index.tsx
