+++ components/ViewMoreCard/index.tsx
